Menu page refactoring
