+++ lib/localization/language.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 import '../main.dart';
-import '../navigation/menu_page.dart';
+import '../pages/menu/menu_page.dart';
 import '../navigation/navigation_helper.dart';
 import 'supported_languages.dart';
